Update handling of plugins
- Load plugins before sendingit to generateTemplateFiles as it's being used for files-processor
- Rename RawPluginConfig to RawPlugin as it better capture what the interface is
