+++ packages/core/src/utils/pluginHelpers/pluginHelpers.ts
@@ -1,7 +1,7 @@
 /**
  * Raw plugin config from reading yml
  */
-export interface RawPluginConfig {
+export interface RawPlugin {
   name: string;
   type?: string;
   config?: Record<string, unknown>;
@@ -17,8 +17,24 @@ export interface CodegenPlugin<C, R> {
 
 export interface PluginModule<C = unknown, R = unknown> {
   plugin: CodegenPlugin<C, R>;
-  config: RawPluginConfig["config"];
+  config: RawPlugin["config"];
 }
+
+// TODO: handle this better e.g. import from node_modules?
+const resolvePluginPath = (pluginName: string): string => {
+  return `../../plugins/${pluginName}`;
+};
+
+const loadPluginModules = async (plugins: RawPlugin[]): Promise<PluginModule[]> => {
+  return await Promise.all<PluginModule>(
+    plugins.map(async (plugin) => {
+      return {
+        plugin: (await import(resolvePluginPath(plugin.name))).plugin,
+        config: plugin.config,
+      };
+    })
+  );
+};
 
 const findFirstOfType = (pluginModules: PluginModule[], type: PluginConfigType): PluginModule | undefined => {
   return pluginModules.find(({ plugin }) => plugin.type === type);
@@ -28,4 +44,4 @@ const filterByTypes = (pluginModules: PluginModule[], types: PluginConfigType[])
   return pluginModules.filter(({ plugin }) => types.includes(plugin.type));
 };
 
-export const pluginHelpers = { findFirstOfType, filterByTypes };
+export const pluginHelpers = { findFirstOfType, filterByTypes, resolvePluginPath, loadPluginModules };
